backend/requirements.txt
Django>=3.2.13
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
requests==2.28.1
requests_toolbelt==0.10.1
jsonpath==0.82
pymysql==1.0.2
rsa==4.9
faker==15.1.3
celery==5.2.7
redis>=4.6
django-celery-beat==2.4.0
uritemplate==4.1.1
coreapi==2.3.3
gunicorn==20.1.0
drf-yasg==1.21.4
supervisor==4.2.4
django-celery-results==2.5.1
channels==3.0.5
channels_redis==4.2.0
django-redis==5.4.0